+++ source/ProcessManager.Orchestrations/Processes/BRS_023_027/V1/Options/OrchestrationOptions_Brs_023_027_V1.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_023_027.V1.Model;
+namespace Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_023_027.V1.Options;
 
 /// <summary>
 /// Options for the <see cref="Orchestration_Brs_023_027_V1"/>.
